requirements.txt
PyYAML==3.11
argh==0.24.1
docopt==0.6.1
globusonline-transfer-api-client==0.10.16
pathtools==0.1.2
watchdog==0.7.1
